Extract shared computeTotalFlowRate to energy.ts
